requirements.txt
numbers_parser==2.3.11
PyYAML==6.0
setuptools==58.1.0
